+++ packages/cli/src/constants.ts
@@ -1,99 +1,11 @@
-import type { Agent } from "node:https";
 import { createRequire } from "node:module";
 import { AgenticMemory } from "@aigne/agentic-memory";
-import { AIGNEHubChatModel } from "@aigne/aigne-hub";
-import { AnthropicChatModel } from "@aigne/anthropic";
-import { BedrockChatModel } from "@aigne/bedrock";
-import type { LoadableModel } from "@aigne/core/loader/index.js";
-import { DeepSeekChatModel } from "@aigne/deepseek";
 import { DefaultMemory } from "@aigne/default-memory";
-import { GeminiChatModel } from "@aigne/gemini";
-import { OllamaChatModel } from "@aigne/ollama";
-import { OpenRouterChatModel } from "@aigne/open-router";
-import { OpenAIChatModel } from "@aigne/openai";
-import { XAIChatModel } from "@aigne/xai";
-import { NodeHttpHandler, streamCollector } from "@smithy/node-http-handler";
-import { HttpsProxyAgent } from "https-proxy-agent";
-import type { ClientOptions } from "openai";
 
 const require = createRequire(import.meta.url);
 
 export const AIGNE_CLI_VERSION = require("../package.json").version;
 
-const GOOGLE = "google";
-
-export function availableModels(): LoadableModel[] {
-  const proxy = ["HTTPS_PROXY", "https_proxy", "HTTP_PROXY", "http_proxy", "ALL_PROXY", "all_proxy"]
-    .map((i) => process.env[i])
-    .filter(Boolean)[0];
-
-  const httpAgent = proxy ? (new HttpsProxyAgent(proxy) as Agent) : undefined;
-  const clientOptions: ClientOptions = {
-    fetchOptions: {
-      // @ts-ignore
-      agent: httpAgent,
-    },
-  };
-
-  return [
-    {
-      name: OpenAIChatModel.name,
-      apiKeyEnvName: "OPENAI_API_KEY",
-      create: (params) => new OpenAIChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: AnthropicChatModel.name,
-      apiKeyEnvName: "ANTHROPIC_API_KEY",
-      create: (params) => new AnthropicChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: BedrockChatModel.name,
-      apiKeyEnvName: "AWS_ACCESS_KEY_ID",
-      create: (params) =>
-        new BedrockChatModel({
-          ...params,
-          clientOptions: {
-            requestHandler: NodeHttpHandler.create({ httpAgent, httpsAgent: httpAgent }),
-            streamCollector,
-          },
-        }),
-    },
-    {
-      name: DeepSeekChatModel.name,
-      apiKeyEnvName: "DEEPSEEK_API_KEY",
-      create: (params) => new DeepSeekChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: GeminiChatModel.name,
-      apiKeyEnvName: "GEMINI_API_KEY",
-      create: (params) => new GeminiChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: GOOGLE,
-      apiKeyEnvName: "GEMINI_API_KEY",
-      create: (params) => new GeminiChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: OllamaChatModel.name,
-      apiKeyEnvName: "OLLAMA_API_KEY",
-      create: (params) => new OllamaChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: OpenRouterChatModel.name,
-      apiKeyEnvName: "OPEN_ROUTER_API_KEY",
-      create: (params) => new OpenRouterChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: XAIChatModel.name,
-      apiKeyEnvName: "XAI_API_KEY",
-      create: (params) => new XAIChatModel({ ...params, clientOptions }),
-    },
-    {
-      name: AIGNEHubChatModel.name,
-      apiKeyEnvName: "AIGNE_HUB_API_KEY",
-      create: (params) => new AIGNEHubChatModel({ ...params, clientOptions }),
-    },
-  ];
-}
+export { availableModels } from "@aigne/aigne-hub";
 
 export const availableMemories = [DefaultMemory, AgenticMemory];
